requirements.txt
django==5.1.4
openai==1.58.1
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3